move logrocket and sentry setup to a separate util file
