+++ src/index.js
@@ -4,50 +4,11 @@ import {BrowserRouter as Router} from 'react-router-dom'
 import {ThemeContextProvider} from './context/ThemeContext'
 import App from './App'
 import './styles/main.css'
-import axios from 'axios'
 import * as serviceWorker from './serviceWorker'
-import * as Sentry from '@sentry/browser'
-import LogRocket from 'logrocket'
-import setupLogRocketReact from 'logrocket-react'
-import {isUndefined} from 'lodash-es'
+import LogRocketSetup from './utils/LogRocketSetup'
 
-const SENTRY_DSN = process.env.REACT_APP_SENTRY_DSN
-Sentry.init({dsn: `${SENTRY_DSN}`})
-
-// init logrocket sessions only in the prod env and on unignored ip
-const initLogRocket = () => {
-  const LOGROCKET_PROJECT_ID = process.env.REACT_APP_LOGROCKET_PROJECT_ID
-  LogRocket.init(`${LOGROCKET_PROJECT_ID}`)
-  setupLogRocketReact(LogRocket)
-
-  // LogRocket and Sentry
-  LogRocket.getSessionURL((sessionURL) => {
-    Sentry.configureScope((scope) => {
-      scope.setExtra('sessionURL', sessionURL)
-    })
-  })
-}
-
-const logRocketSetup = async () => {
-  const ipaddresses = process.env.REACT_APP_IGNORE_IP
-  // check if ignore ip env variable exist
-  if (!isUndefined(ipaddresses)) {
-    // fetch ip address and check if it matches any one of the ipaddresses
-    // initLogRocket when ip is not in the ignore list or ip fetch fail
-    const {ip} = (await axios.get('https://api.myip.com')).data
-    if (
-      !isUndefined(ip) &&
-      !ipaddresses.split(',').includes(ip) &&
-      process.env.NODE_ENV !== 'development'
-    ) {
-      initLogRocket()
-    }
-  } else {
-    initLogRocket()
-  }
-}
-
-logRocketSetup()
+// setup LogRocket & Sentry
+LogRocketSetup()
 
 // custom console message
 console.clear()
